+++ src/FunFair.BuildCheck.ProjectChecks/ReferencedPackages/MustNotReferenceTeamCityTestAdapter.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.Logging;
+
+namespace FunFair.BuildCheck.ProjectChecks.ReferencedPackages;
+
+public sealed class MustNotReferenceTeamCityTestAdapter : MustNotReferencePackages
+{
+    public MustNotReferenceTeamCityTestAdapter(ILogger<MustNotReferenceTeamCityTestAdapter> logger)
+        : base(["TeamCity.VSTest.TestAdapter"], reason: "No longer using teamcity", logger: logger)
+    {
+    }
+}